refactored code structure directories
