+++ math_trainer/problem/__init__.py
@@ -1,0 +1,2 @@
+from math_trainer.problem.utils import *
+from math_trainer.problem.problems import *
