+++ math_trainer/utils.py
@@ -1,5 +1,5 @@
 import os
-from typing import Union, List, Any
+from typing import Any, List, Union
 
 import numpy as np
 import pandas as pd
